Extract the create file and remove file
