+++ utils/utils.go
@@ -6,6 +6,8 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+	"os"
+	"path/filepath"
 	"strings"
 
 	"github.com/joho/godotenv"
@@ -62,4 +64,44 @@ func WriteJSON(w http.ResponseWriter, status int, v any) {
 	if err := json.NewEncoder(w).Encode(v); err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 	}
+}
+
+// EnsureDirectoryAndFile checks if a directory and file exist and creates them if they don't
+func EnsureDirectoryAndFile(path, filename string) (string, error) {
+	// Check if directory exists
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		fmt.Println("Directory does not exist, creating it...")
+		if err := os.MkdirAll(path, os.ModePerm); err != nil {
+			return "", fmt.Errorf("failed to create directory: %w", err)
+		}
+	}
+
+	// Combine directory and filename to get the full file path
+	filePath := filepath.Join(path, filename)
+
+	// Check if the file exists in the directory
+	if _, err := os.Stat(filePath); os.IsNotExist(err) {
+		fmt.Println("File does not exist, creating it...")
+		file, err := os.Create(filePath)
+		if err != nil {
+			return "", fmt.Errorf("failed to create file: %w", err)
+		}
+		defer file.Close()
+	}
+
+	return filePath, nil
+}
+
+// Remove File
+func RemoveFile(filepath string) error {
+	if _, err := os.Stat(filepath); os.IsNotExist(err) {
+		fmt.Println("Nothing to delete")
+		return fmt.Errorf("File %s doesn't exist", err)
+	}
+
+	if err := os.Remove(filepath); err != nil {
+		return err
+	}
+
+	return nil
 }
